+++ apps/cli/cli_main.c
@@ -439,6 +439,9 @@ main(int argc, char **argv)
     argc -= optind;
     argv += optind;
 
+    /* Access the remaining argv/argc options (after --) w clicon-argv_get() */
+    clicon_argv_set(h, argv0, argc, argv);
+
     /* Defer: Wait to the last minute to print help message */
     if (help)
 	usage(h, argv[0]);